+++ .gitignore
@@ -29,4 +29,3 @@ replay_pid*
 # Ignore Gradle build output directory
 build
 
-.vscode
