--- app/src/main/java/be/craftcode/ap/collections/Java21Collections.java
@@ -1,5 +1,0 @@
-package be.craftcode.ap.collections;
-
-public class Java21Collections {
-    
-}
