+++ app/src/main/java/be/craftcode/ap/java21Lab/App.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package be.craftcode.ap;
+package be.craftcode.ap.java21Lab;
 
 public class App {
     public String getGreeting() {
