+++ app/src/main/java/be/craftcode/ap/java21Lab/collections/Java21Collections.java
@@ -1,0 +1,5 @@
+package be.craftcode.ap.java21Lab.collections;
+
+public class Java21Collections {
+    
+}
