+++ app/src/test/java/be/craftcode/ap/java21Lab/AppTest.java
@@ -1,9 +1,10 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package be.craftcode.ap;
+package be.craftcode.ap.java21Lab;
 
 import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class AppTest {
