+++ app/src/test/java/be/craftcode/ap/java21Lab/collections/Java21CollectionsTest.java
@@ -1,0 +1,8 @@
+package be.craftcode.ap.java21Lab.collections;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class Java21CollectionsTest {
+}
